Rename interfaces
