+++ src/Contracts/CurrencyFactoriesLocator.php
@@ -4,7 +4,7 @@ namespace Adsmurai\Currency\Contracts;
 
 interface CurrencyFactoriesLocator
 {
-    public function getCurrencyFactory(string $isoCode): CurrencyFactory;
+    public function getCurrencyFactory(string $isoCode): MoneyFactory;
 
     /** @return string[] */
     public function getSupportedCurrencyISOCodes(): array;
