+++ src/Contracts/CurrencyFactory.php
@@ -2,15 +2,10 @@
 
 namespace Adsmurai\Currency\Contracts;
 
-use Litipk\BigNumbers\Decimal;
-
 interface CurrencyFactory
 {
-    public function buildFromFloat(float $amount): Currency;
+    public function buildFromISOCode(string $ISOCode): Currency;
 
-    public function buildFromFractionalUnits(int $amount): Currency;
-
-    public function buildFromString(string $amount): Currency;
-
-    public function buildFromDecimal(Decimal $amount): Currency;
+    /** @return string[] */
+    public function getSupportedCurrencyISOCodes(): array;
 }
