--- src/Contracts/CurrencyTypeFactory.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Adsmurai\Currency\Contracts;
-
-interface CurrencyTypeFactory
-{
-    public function buildFromISOCode(string $ISOCode): CurrencyType;
-
-    /** @return string[] */
-    public function getSupportedCurrencyISOCodes(): array;
-}
