+++ src/Errors/InvalidCurrenciesDataError.php
@@ -5,7 +5,7 @@ namespace Adsmurai\Currency\Errors;
 use Adsmurai\Currency\Contracts\CurrencyError;
 use LogicException;
 
-final class InvalidCurrencyTypesDataError extends LogicException implements CurrencyError
+final class InvalidCurrenciesDataError extends LogicException implements CurrencyError
 {
     public function __construct()
     {
