+++ tests/Currency/CurrencyTypeMocks.php
@@ -4,31 +4,31 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\Currency;
 
-use Adsmurai\Currency\Contracts\CurrencyType;
+use Adsmurai\Currency\Contracts\Currency;
 use Mockery\MockInterface;
 
 final class CurrencyTypeMocks
 {
-    /**  @return CurrencyType|MockInterface */
-    public static function getComparableTwoDecimalDigitsCurrencyType(bool $equals = true): CurrencyType
+    /**  @return Currency|MockInterface */
+    public static function getComparableTwoDecimalDigitsCurrencyType(bool $equals = true): Currency
     {
-        /** @var CurrencyType|MockInterface $currencyType */
+        /** @var Currency|MockInterface $currencyType */
         $currencyType = self::getTwoDecimalDigitsCurrencyType();
         $currencyType->shouldReceive('equals')->andReturn($equals);
 
         return $currencyType;
     }
 
-    /**  @return CurrencyType|MockInterface */
-    public static function getTwoDecimalDigitsCurrencyType(): CurrencyType
+    /**  @return Currency|MockInterface */
+    public static function getTwoDecimalDigitsCurrencyType(): Currency
     {
         return self::getNDecimalDigitsCurrencyType(2);
     }
 
-    /**  @return CurrencyType|MockInterface */
-    public static function getNDecimalDigitsCurrencyType(int $n = 2): CurrencyType
+    /**  @return Currency|MockInterface */
+    public static function getNDecimalDigitsCurrencyType(int $n = 2): Currency
     {
-        /** @var CurrencyType|MockInterface $currencyType */
+        /** @var Currency|MockInterface $currencyType */
         $currencyType = self::getCurrencyTypeDummyMock();
         $currencyType
             ->shouldReceive('getNumFractionalDigits')->andReturn($n);
@@ -36,39 +36,39 @@ final class CurrencyTypeMocks
         return $currencyType;
     }
 
-    /**  @return CurrencyType|MockInterface */
-    public static function getCurrencyTypeDummyMock(): CurrencyType
+    /**  @return Currency|MockInterface */
+    public static function getCurrencyTypeDummyMock(): Currency
     {
-        /** @var CurrencyType|MockInterface $currencyType */
-        $currencyType = \Mockery::mock(CurrencyType::class);
+        /** @var Currency|MockInterface $currencyType */
+        $currencyType = \Mockery::mock(Currency::class);
 
         return $currencyType;
     }
 
-    public static function getEuroCurrencyType(): CurrencyType
+    public static function getEuroCurrencyType(): Currency
     {
-        /** @var CurrencyType|MockInterface $currencyType */
-        $currencyType = \Mockery::mock(CurrencyType::class);
+        /** @var Currency|MockInterface $currencyType */
+        $currencyType = \Mockery::mock(Currency::class);
 
         $currencyType
             ->shouldReceive('getNumFractionalDigits')->andReturn(2)
             ->shouldReceive('getISOCode')->andReturn('EUR')
             ->shouldReceive('getSymbol')->andReturn('€')
-            ->shouldReceive('getSymbolPlacement')->andReturn(CurrencyType::AFTER_PLACEMENT);
+            ->shouldReceive('getSymbolPlacement')->andReturn(Currency::AFTER_PLACEMENT);
 
         return $currencyType;
     }
 
-    public static function getUsDollarCurrencyType(): CurrencyType
+    public static function getUsDollarCurrencyType(): Currency
     {
-        /** @var CurrencyType|MockInterface $currencyType */
-        $currencyType = \Mockery::mock(CurrencyType::class);
+        /** @var Currency|MockInterface $currencyType */
+        $currencyType = \Mockery::mock(Currency::class);
 
         $currencyType
             ->shouldReceive('getNumFractionalDigits')->andReturn(2)
             ->shouldReceive('getISOCode')->andReturn('USD')
             ->shouldReceive('getSymbol')->andReturn('$')
-            ->shouldReceive('getSymbolPlacement')->andReturn(CurrencyType::BEFORE_PLACEMENT);
+            ->shouldReceive('getSymbolPlacement')->andReturn(Currency::BEFORE_PLACEMENT);
 
         return $currencyType;
     }
