+++ tests/Currency/FormatTests.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\Currency;
 
-use Adsmurai\Currency\Contracts\CurrencyType;
-use Adsmurai\Currency\Currency;
-use Adsmurai\Currency\CurrencyFormat;
+use Adsmurai\Currency\Contracts\Currency;
+use Adsmurai\Currency\Money;
+use Adsmurai\Currency\MoneyFormat;
 use Mockery\MockInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -14,115 +14,121 @@ class formatTests extends TestCase
 {
     /**
      * @dataProvider simpleParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_without_custom_parameters(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         string $formattedCurrency
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
+        $currency = Money::fromString($amount, $currencyType);
         $this->assertEquals($formattedCurrency, $currency->format());
     }
 
     /**
      * @dataProvider customizedParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_custom_separators(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         string $decimalsSeparator,
         string $thousandsSeparator,
         string $formattedCurrency
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParameters($decimalsSeparator, $thousandsSeparator);
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParameters($decimalsSeparator, $thousandsSeparator);
         $this->assertEquals($formattedCurrency, $currency->format($currencyFormat));
     }
 
     /**
      * @dataProvider extraPrecisionParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_extra_precision(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         int $extraPrecision,
         string $formattedCurrency
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParametersWithExtraPrecision($extraPrecision);
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParametersWithExtraPrecision($extraPrecision);
         $this->assertEquals($formattedCurrency, $currency->format($currencyFormat));
     }
 
     /**
      * @dataProvider precisionParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_custom_precision(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         int $precision,
         string $formattedCurrency
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParametersWithPrecision($precision);
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParametersWithPrecision($precision);
         $this->assertEquals($formattedCurrency, $currency->format($currencyFormat));
     }
 
     /**
      * @dataProvider noDecorationParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_no_decoration(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         string $formattedCurrency
     ) {
-        //$this->markTestSkipped();
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParameters(
-            CurrencyFormat::DEFAULT_DECIMALS_SEPARATOR,
-            CurrencyFormat::DEFAULT_THOUSANDS_SEPARATOR,
-            CurrencyFormat::DECORATION_NO_DECORATION
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParameters(
+            MoneyFormat::DEFAULT_DECIMALS_SEPARATOR,
+            MoneyFormat::DEFAULT_THOUSANDS_SEPARATOR,
+            MoneyFormat::DECORATION_NO_DECORATION
         );
         $this->assertEquals($formattedCurrency, $currency->format($currencyFormat));
     }
 
     /**
      * @dataProvider isoCodeDecorationParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_decoration_iso_code(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         string $formattedCurrency
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParameters(
-            CurrencyFormat::DEFAULT_DECIMALS_SEPARATOR,
-            CurrencyFormat::DEFAULT_THOUSANDS_SEPARATOR,
-            CurrencyFormat::DECORATION_ISO_CODE
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParameters(
+            MoneyFormat::DEFAULT_DECIMALS_SEPARATOR,
+            MoneyFormat::DEFAULT_THOUSANDS_SEPARATOR,
+            MoneyFormat::DECORATION_ISO_CODE
         );
         $this->assertEquals($formattedCurrency, $currency->format($currencyFormat));
     }
 
     /**
      * @dataProvider spaceDecorationParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::format
+     * @covers       \Adsmurai\Currency\Money::format
+     * @covers       \Adsmurai\Currency\Money::decorate
      */
     public function test_format_with_separator(
         string $amount,
-        CurrencyType $currencyType,
+        Currency $currencyType,
         string $formattedCurrency,
         int $decorationSpace,
         int $decorationType
     ) {
-        $currency = Currency::fromString($amount, $currencyType);
-        $currencyFormat = CurrencyFormat::fromParameters(
-            CurrencyFormat::DEFAULT_DECIMALS_SEPARATOR,
-            CurrencyFormat::DEFAULT_THOUSANDS_SEPARATOR,
+        $currency = Money::fromString($amount, $currencyType);
+        $currencyFormat = MoneyFormat::fromParameters(
+            MoneyFormat::DEFAULT_DECIMALS_SEPARATOR,
+            MoneyFormat::DEFAULT_THOUSANDS_SEPARATOR,
             $decorationType,
             $decorationSpace
         );
@@ -147,10 +153,10 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$'), '0.010$'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$'), '12345678.500$'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '$34.760'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '$100.000'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '$0.010'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '$12345678.500'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '$34.760'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '$100.000'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '$0.010'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '$12345678.500'],
 
             ['34.761', $this->getNDecimalDigitsCurrencyType(), '34.76€'],
             ['34.766', $this->getNDecimalDigitsCurrencyType(), '34.77€'],
@@ -166,11 +172,11 @@ class formatTests extends TestCase
     private static function getNDecimalDigitsCurrencyType(
         int $n = 2,
         string $symbol = '€',
-        int $symbolPlacement = CurrencyType::AFTER_PLACEMENT,
+        int $symbolPlacement = Currency::AFTER_PLACEMENT,
         string $isoCode = 'EUR'
-    ): CurrencyType {
-        /** @var CurrencyType|MockInterface $currencyType */
-        $currencyType = \Mockery::mock(CurrencyType::class);
+    ): Currency {
+        /** @var Currency|MockInterface $currencyType */
+        $currencyType = \Mockery::mock(Currency::class);
 
         $currencyType
             ->shouldReceive('getNumFractionalDigits')->andReturn($n)
@@ -199,10 +205,10 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$'), ',', '.', '0,010$'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$'), ',', '.', '12.345.678,500$'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), ',', '.', '$34,760'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), ',', '.', '$100,000'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), ',', '.', '$0,010'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), ',', '.', '$12.345.678,500'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), ',', '.', '$34,760'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), ',', '.', '$100,000'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), ',', '.', '$0,010'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), ',', '.', '$12.345.678,500'],
         ];
     }
 
@@ -224,10 +230,10 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$'), 2, '0.01000$'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$'), 2, '12345678.50000$'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$34.76000'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$100.00000'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$0.01000'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$12345678.50000'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$34.76000'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$100.00000'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$0.01000'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$12345678.50000'],
         ];
     }
 
@@ -249,10 +255,10 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$'), 2, '0.01$'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$'), 2, '12345678.50$'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$34.76'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$100.00'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$0.01'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), 2, '$12345678.50'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$34.76'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$100.00'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$0.01'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), 2, '$12345678.50'],
         ];
     }
 
@@ -274,10 +280,10 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$'), '0.010'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$'), '12345678.500'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '34.760'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '100.000'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '0.010'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT), '12345678.500'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '34.760'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '100.000'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '0.010'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT), '12345678.500'],
 
             ['34.761', $this->getNDecimalDigitsCurrencyType(), '34.76'],
             ['34.766', $this->getNDecimalDigitsCurrencyType(), '34.77'],
@@ -303,15 +309,15 @@ class formatTests extends TestCase
             ['0.01', $this->getNDecimalDigitsCurrencyType(3), '0.010EUR'],
             ['12345678.50', $this->getNDecimalDigitsCurrencyType(3), '12345678.500EUR'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::AFTER_PLACEMENT, 'USD'), '34.760USD'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::AFTER_PLACEMENT, 'USD'), '100.000USD'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::AFTER_PLACEMENT, 'USD'), '0.010USD'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::AFTER_PLACEMENT, 'USD'), '12345678.500USD'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::AFTER_PLACEMENT, 'USD'), '34.760USD'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::AFTER_PLACEMENT, 'USD'), '100.000USD'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::AFTER_PLACEMENT, 'USD'), '0.010USD'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::AFTER_PLACEMENT, 'USD'), '12345678.500USD'],
 
-            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT, 'USD'), '34.760USD'],
-            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT, 'USD'), '100.000USD'],
-            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT, 'USD'), '0.010USD'],
-            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', CurrencyType::BEFORE_PLACEMENT, 'USD'), '12345678.500USD'],
+            ['34.76', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT, 'USD'), '34.760USD'],
+            ['100', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT, 'USD'), '100.000USD'],
+            ['0.01', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT, 'USD'), '0.010USD'],
+            ['12345678.50', $this->getNDecimalDigitsCurrencyType(3, '$', Currency::BEFORE_PLACEMENT, 'USD'), '12345678.500USD'],
 
             ['34.761', $this->getNDecimalDigitsCurrencyType(), '34.76EUR'],
             ['34.766', $this->getNDecimalDigitsCurrencyType(), '34.77EUR'],
@@ -331,43 +337,43 @@ class formatTests extends TestCase
                 '34.76',
                 $this->getNDecimalDigitsCurrencyType(),
                 '34.76 €',
-                CurrencyFormat::DECORATION_WITH_SPACE,
-                CurrencyFormat::DECORATION_SYMBOL,
+                MoneyFormat::DECORATION_WITH_SPACE,
+                MoneyFormat::DECORATION_SYMBOL,
             ],
             [
                 '34.76',
                 $this->getNDecimalDigitsCurrencyType(),
                 '34.76€',
-                CurrencyFormat::DECORATION_WITHOUT_SPACE,
-                CurrencyFormat::DECORATION_SYMBOL,
+                MoneyFormat::DECORATION_WITHOUT_SPACE,
+                MoneyFormat::DECORATION_SYMBOL,
             ],
             [
                 '34.76',
                 $this->getNDecimalDigitsCurrencyType(),
                 '34.76 EUR',
-                CurrencyFormat::DECORATION_WITH_SPACE,
-                CurrencyFormat::DECORATION_ISO_CODE,
+                MoneyFormat::DECORATION_WITH_SPACE,
+                MoneyFormat::DECORATION_ISO_CODE,
             ],
             [
                 '34.76',
                 $this->getNDecimalDigitsCurrencyType(),
                 '34.76EUR',
-                CurrencyFormat::DECORATION_WITHOUT_SPACE,
-                CurrencyFormat::DECORATION_ISO_CODE,
+                MoneyFormat::DECORATION_WITHOUT_SPACE,
+                MoneyFormat::DECORATION_ISO_CODE,
             ],
             [
                 '34.76',
-                $this->getNDecimalDigitsCurrencyType(2, '€', CurrencyType::BEFORE_PLACEMENT),
+                $this->getNDecimalDigitsCurrencyType(2, '€', Currency::BEFORE_PLACEMENT),
                 '€ 34.76',
-                CurrencyFormat::DECORATION_WITH_SPACE,
-                CurrencyFormat::DECORATION_SYMBOL,
+                MoneyFormat::DECORATION_WITH_SPACE,
+                MoneyFormat::DECORATION_SYMBOL,
             ],
             [
                 '34.76',
-                $this->getNDecimalDigitsCurrencyType(2, '€', CurrencyType::BEFORE_PLACEMENT),
+                $this->getNDecimalDigitsCurrencyType(2, '€', Currency::BEFORE_PLACEMENT),
                 '€34.76',
-                CurrencyFormat::DECORATION_WITHOUT_SPACE,
-                CurrencyFormat::DECORATION_SYMBOL,
+                MoneyFormat::DECORATION_WITHOUT_SPACE,
+                MoneyFormat::DECORATION_SYMBOL,
             ],
         ];
     }
