+++ tests/Currency/FromFloatTests.php
@@ -4,59 +4,59 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\Currency;
 
-use Adsmurai\Currency\Contracts\Currency as CurrencyInterface;
-use Adsmurai\Currency\Contracts\CurrencyType;
-use Adsmurai\Currency\Currency;
+use Adsmurai\Currency\Contracts\Money as CurrencyInterface;
+use Adsmurai\Currency\Contracts\Currency;
+use Adsmurai\Currency\Money;
 use PHPUnit\Framework\TestCase;
 
 class fromFloatTests extends TestCase
 {
     /**
      * @dataProvider validParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::fromFloat
-     * @covers       \Adsmurai\Currency\Currency::__construct
+     * @covers       \Adsmurai\Currency\Money::fromFloat
+     * @covers       \Adsmurai\Currency\Money::__construct
      */
-    public function test_with_valid_params(float $amount, CurrencyType $currencyType)
+    public function test_with_valid_params(float $amount, Currency $currencyType)
     {
-        $currency = Currency::fromFloat($amount, $currencyType);
+        $currency = Money::fromFloat($amount, $currencyType);
 
         $this->assertInstanceOf(CurrencyInterface::class, $currency);
-        $this->assertInstanceOf(Currency::class, $currency);
+        $this->assertInstanceOf(Money::class, $currency);
     }
 
     /**
      * @dataProvider negativeParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::fromFloat
-     * @covers       \Adsmurai\Currency\Currency::__construct
+     * @covers       \Adsmurai\Currency\Money::fromFloat
+     * @covers       \Adsmurai\Currency\Money::__construct
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessage Currency amounts must be positive
      */
-    public function test_with_negative_amounts(float $amount, CurrencyType $currencyType)
+    public function test_with_negative_amounts(float $amount, Currency $currencyType)
     {
-        Currency::fromFloat($amount, $currencyType);
+        Money::fromFloat($amount, $currencyType);
     }
 
     /**
      * @dataProvider infiniteParamsProvider
-     * @covers       \Adsmurai\Currency\Currency::fromFloat
-     * @covers       \Adsmurai\Currency\Currency::__construct
+     * @covers       \Adsmurai\Currency\Money::fromFloat
+     * @covers       \Adsmurai\Currency\Money::__construct
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessage Currency amounts must be finite
      */
-    public function test_with_infinite_amounts(float $amount, CurrencyType $currencyType)
+    public function test_with_infinite_amounts(float $amount, Currency $currencyType)
     {
-        Currency::fromFloat($amount, $currencyType);
+        Money::fromFloat($amount, $currencyType);
     }
 
     /**
-     * @covers \Adsmurai\Currency\Currency::fromFloat
-     * @covers \Adsmurai\Currency\Currency::__construct
+     * @covers \Adsmurai\Currency\Money::fromFloat
+     * @covers \Adsmurai\Currency\Money::__construct
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessage Currency amounts must be numbers
      */
     public function test_with_nan_amount()
     {
-        Currency::fromFloat(\NAN, CurrencyTypeMocks::getCurrencyTypeDummyMock());
+        Money::fromFloat(\NAN, CurrencyTypeMocks::getCurrencyTypeDummyMock());
     }
 
     public function validParamsProvider(): array
