+++ tests/CurrencyFactoriesLocator/GetSupportedCurrencyISOCodesTests.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\CurrencyFactoriesLocator;
 
-use Adsmurai\Currency\Contracts\CurrencyTypeFactory;
+use Adsmurai\Currency\Contracts\CurrencyFactory;
 use Adsmurai\Currency\CurrencyFactoriesLocator;
 use PHPUnit\Framework\TestCase;
 
@@ -15,8 +15,8 @@ class GetSupportedCurrencyISOCodesTests extends TestCase
      */
     public function test_that_returns_the_supported_currency_iso_codes()
     {
-        /** @var \Mockery\MockInterface|CurrencyTypeFactory $currencyTypeFactory */
-        $currencyTypeFactory = \Mockery::mock(CurrencyTypeFactory::class);
+        /** @var \Mockery\MockInterface|CurrencyFactory $currencyTypeFactory */
+        $currencyTypeFactory = \Mockery::mock(CurrencyFactory::class);
         $currencyTypeFactory->shouldReceive('getSupportedCurrencyISOCodes')->andReturn(['EUR', 'USD']);
 
         $currencyFactoriesLocator = new CurrencyFactoriesLocator($currencyTypeFactory);
