+++ tests/CurrencyTypeFactory/BuildFromISOCodeTests.php
@@ -4,25 +4,25 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\CurrencyTypeFactory;
 
-use Adsmurai\Currency\Contracts\CurrencyType;
-use Adsmurai\Currency\CurrencyTypeFactory;
+use Adsmurai\Currency\Contracts\Currency;
+use Adsmurai\Currency\CurrencyFactory;
 use PHPUnit\Framework\TestCase;
 
 class BuildFromISOCodeTests extends TestCase
 {
     /**
      * @dataProvider commonCurrenciesProvider
-     * @covers       \Adsmurai\Currency\CurrencyTypeFactory
+     * @covers       \Adsmurai\Currency\CurrencyFactory
      */
     public function test_with_common_currencies(string $ISOCode)
     {
         /** @var array $currencyData */
         $currencyData = $this->getCurrenciesData()[$ISOCode];
-        $currencyTypeFactory = CurrencyTypeFactory::fromDataPath();
+        $currencyTypeFactory = CurrencyFactory::fromDataPath();
 
         $currencyType = $currencyTypeFactory->buildFromISOCode($ISOCode);
 
-        $this->assertInstanceOf(CurrencyType::class, $currencyType);
+        $this->assertInstanceOf(Currency::class, $currencyType);
         $this->assertEquals($ISOCode, $currencyType->getISOCode());
         $this->assertEquals($currencyData['symbol'], $currencyType->getSymbol());
         $this->assertEquals($currencyData['symbolPlacement'], $currencyType->getSymbolPlacement());
@@ -39,11 +39,11 @@ class BuildFromISOCodeTests extends TestCase
 
     /**
      * @dataProvider commonCurrenciesProvider
-     * @covers       \Adsmurai\Currency\CurrencyTypeFactory
+     * @covers       \Adsmurai\Currency\CurrencyFactory
      */
     public function test_that_there_are_no_multiple_instances_for_same_currency_type(string $ISOCode)
     {
-        $currencyTypeFactory = CurrencyTypeFactory::fromDataPath();
+        $currencyTypeFactory = CurrencyFactory::fromDataPath();
 
         $currencyTypeA = $currencyTypeFactory->buildFromISOCode($ISOCode);
         $currencyTypeB = $currencyTypeFactory->buildFromISOCode($ISOCode);
@@ -52,16 +52,16 @@ class BuildFromISOCodeTests extends TestCase
     }
 
     /**
-     * @covers \Adsmurai\Currency\CurrencyTypeFactory::buildFromISOCode
+     * @covers \Adsmurai\Currency\CurrencyFactory::buildFromISOCode
      * @expectedException \Adsmurai\Currency\Errors\UnsupportedCurrencyISOCodeError
      * @expectedExceptionMessage Unsupported currency ISO code (USD)
      */
     public function test_that_an_exception_is_thrown_when_we_try_to_build_a_not_defined_currency_type()
     {
-        $currencyTypeFactory = CurrencyTypeFactory::fromDataArray(['EUR' => [
+        $currencyTypeFactory = CurrencyFactory::fromDataArray(['EUR' => [
             'numFractionalDigits' => 2,
             'symbol' => '€',
-            'symbolPlacement' => CurrencyType::AFTER_PLACEMENT,
+            'symbolPlacement' => Currency::AFTER_PLACEMENT,
         ]]);
 
         $currencyTypeFactory->buildFromISOCode('USD');
