+++ tests/CurrencyTypeFactory/GetSupportedCurrencyISOCodesTests.php
@@ -4,27 +4,27 @@ declare(strict_types=1);
 
 namespace Adsmurai\Currency\Tests\CurrencyTypeFactory;
 
-use Adsmurai\Currency\Contracts\CurrencyType;
-use Adsmurai\Currency\CurrencyTypeFactory;
+use Adsmurai\Currency\Contracts\Currency;
+use Adsmurai\Currency\CurrencyFactory;
 use PHPUnit\Framework\TestCase;
 
 class GetSupportedCurrencyISOCodesTests extends TestCase
 {
     /**
-     * @covers \Adsmurai\Currency\CurrencyTypeFactory::getSupportedCurrencyISOCodes
+     * @covers \Adsmurai\Currency\CurrencyFactory::getSupportedCurrencyISOCodes
      */
     public function test_that_returns_the_supported_currency_iso_codes()
     {
-        $currencyTypeFactory = CurrencyTypeFactory::fromDataArray([
+        $currencyTypeFactory = CurrencyFactory::fromDataArray([
             'EUR' => [
                 'numFractionalDigits' => 2,
                 'symbol' => '€',
-                'symbolPlacement' => CurrencyType::AFTER_PLACEMENT,
+                'symbolPlacement' => Currency::AFTER_PLACEMENT,
             ],
             'USD' => [
                 'numFractionalDigits' => 2,
                 'symbol' => '$',
-                'symbolPlacement' => CurrencyType::BEFORE_PLACEMENT,
+                'symbolPlacement' => Currency::BEFORE_PLACEMENT,
             ],
         ]);
